Add random cropping while training

+++ data/unaligned_dataset.py
@@ -59,7 +59,7 @@ class UnalignedDataset(BaseDataset):
         A_mask_path = A_path.replace("trainA", "maskA")
         A_img = Image.open(A_path).convert('RGB')
         B_img = Image.open(B_path).convert('RGB')
-        A_mask_image = (np.array(Image.open(A_mask_path)) > 0).astype(np.int64)
+        A_mask_image = (np.array(Image.open(A_mask_path)) > 0).astype(np.float64)
         # apply image transformation
         A = self.transform_A(A_img)
         B = self.transform_B(B_img)
